+++ xtreme-slider.php
@@ -1,9 +1,9 @@
 <?php
 /**
- * Plugin Name: XtremeSlider
+ * Plugin Name: Xtreme Slider (3D Image Slider)
  * Plugin URI: https://xtremeplugins.com/plugins/xtreme-slider/
  * Description: Beautiful image slider with Simple and 3D layouts. Upload photos, generate shortcodes, and embed anywhere in WordPress or Elementor.
- * Version: 1.4.1
+ * Version: 1.4.3
  * Author: XtremePlugins
  * Author URI: https://xtremeplugins.com
  * License: GPL-2.0-or-later
@@ -16,7 +16,7 @@
 
 defined( 'ABSPATH' ) || exit;
 
-define( 'XS_VERSION',         '1.4.1' );
+define( 'XS_VERSION',         '1.4.3' );
 define( 'XS_DB_VERSION',      '8' );
 define( 'XS_PLUGIN_FILE',     __FILE__ );
 define( 'XS_PLUGIN_PATH',     plugin_dir_path( __FILE__ ) );